move components to right folder, modify onclick props, add export button
